backend/requirements.txt
fastapi==0.120.2
uvicorn==0.38.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-multipart==0.0.12

passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
email-validator==2.3.0

